+++ apps/backend/clixon_backend_handle.c
@@ -44,7 +44,6 @@
 #include <string.h>
 #include <stdlib.h>
 #include <inttypes.h>
-#include <dirent.h>
 #include <errno.h>
 #include <unistd.h>
 #include <assert.h>