+++ apps/backend/clixon_backend_transaction.c
@@ -46,7 +46,6 @@
 #include <string.h>
 #include <stdlib.h>
 #include <inttypes.h>
-#include <dirent.h>
 #include <errno.h>
 #include <ctype.h>
 #include <sys/types.h>