+++ test/test_perf_restconf.sh
@@ -150,7 +150,7 @@ sed '/Connection:/d' $foutput > $foutput2 && mv $foutput2 $foutput
 
 # Create a file to compare with
 if ${HAVE_LIBNGHTTP2}; then
-    if [ ${HAVE_LIBEVHTP} -a ${RCPROTO} = http ]; then
+    if [ ${HAVE_HTTP1} -a ${RCPROTO} = http ]; then
 	# Add 101 switch protocols for http 1->2 upgrade
 	echo "HTTP/1.1 101 Switching Protocols" > $ftest
         echo "Upgrade: h2c" >> $ftest