+++ test/test_restconf_basic_auth.sh
@@ -14,8 +14,8 @@
 # Magic line must be first in script (see README.md)
 s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi
 
-if ! ${HAVE_LIBEVHTP}; then
-    echo "...skipped: LIBEVHTP is false, must run with http/1 (evhtp)"
+if ! ${HAVE_HTTP1}; then
+    echo "...skipped: Must run with http/1"
     if [ "$s" = $0 ]; then exit 0; else return 0; fi
 fi