+++ test/test_restconf_internal_usecases.sh
@@ -25,9 +25,9 @@
 # Magic line must be first in script (see README.md)
 s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi
 
-# Does not work with evhtpnative http/2-only
-if [ "${WITH_RESTCONF}" = "native" -a ${HAVE_LIBEVHTP} = false ]; then
-    echo "...skipped: LIBEVHTP is false, must run with http/1 (evhtp)"
+# Does not work with native http/2-only
+if [ "${WITH_RESTCONF}" = "native" -a ${HAVE_HTTP1} = false ]; then
+    echo "...skipped: Must run with http/1"
     if [ "$s" = $0 ]; then exit 0; else return 0; fi
 fi